Fixed all build errors.
