+++ Snippets/SqlServer/SqlServer_2.1/ConnectionStrings.xml
@@ -1,0 +1,9 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<!--startcode sqlserver-connection-string-xml-->
+<configuration>
+  <connectionStrings>
+     <add name="NServiceBus/Transport"
+          connectionString="Data Source=INSTANCE_NAME; Initial Catalog=some_database; Integrated Security=True"/>
+  </connectionStrings>
+</configuration>
+<!--endcode-->
