+++ Snippets/SqlServer/SqlServer_2.1/ConnectionStringsNamed.xml
@@ -1,0 +1,9 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<!--startcode sqlserver-named-connection-string-xml-->
+<configuration>
+  <connectionStrings>
+    <add name="MyConnectionString"
+         connectionString="Data Source=INSTANCE_NAME; Initial Catalog=some_database; Integrated Security=True; Queue Schema=nsb"/>
+  </connectionStrings>
+</configuration>
+<!--endcode-->
