+++ Snippets/SqlServer/SqlServer_2.1/ExpiredIndex.sql
@@ -1,0 +1,12 @@
+﻿--startcode sql-ExpiresIndex
+CREATE NONCLUSTERED INDEX [Index_Expires]
+ON [schema].[queuename]
+(
+	[Expires] ASC
+)
+INCLUDE
+(
+	[Id],
+	[RowVersion]
+)
+--endcode
