+++ Snippets/SqlServer/SqlServer_2.1/MultiDb.config.xml
@@ -1,0 +1,21 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+  <!--startcode sqlserver-multidb-connectionstrings-->
+  <connectionStrings>
+    <add name="NServiceBus/Transport"
+         connectionString="Server=DbServerA;Database=MyDefaultDB;"/>
+    <add name="NServiceBus/Transport/Billing"
+         connectionString="Server=DbServerB;Database=Billing;"/>
+  </connectionStrings>
+  <!--endcode-->
+  <!--startcode sqlserver-multidb-messagemapping-->
+  <UnicastBusConfig>
+    <MessageEndpointMappings>
+      <add Assembly="Billing.Contract"
+           Endpoint="billing"/>
+      <add Assembly="Sales.Contract"
+           Endpoint="sales"/>
+    </MessageEndpointMappings>
+  </UnicastBusConfig>
+  <!--endcode-->
+</configuration>
