+++ Snippets/SqlServer/SqlServer_2.1/MultiSchema.xml
@@ -1,0 +1,25 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+  <!--startcode sqlserver-non-standard-schema-connString-xml -->
+  <connectionStrings>
+    <add name="NServiceBus/Transport"
+         connectionString="Data Source=INSTANCE_NAME; Initial Catalog=some_database; Integrated Security=True; Queue Schema=myschema"/>
+  </connectionStrings>
+  <!--endcode-->
+  <!--startcode sqlserver-non-standard-schema-messagemapping -->
+  <connectionStrings>
+    <add name="NServiceBus/Transport"
+         connectionString="Data Source=INSTANCE_NAME; Initial Catalog=some_database; Integrated Security=True;Queue Schema=salesSchema"/>
+    <add name="NServiceBus/Transport/Billing"
+         connectionString="Data Source=INSTANCE_NAME; Initial Catalog=some_database; Integrated Security=True;Queue Schema=billingSchema"/>
+  </connectionStrings>
+  <UnicastBusConfig>
+    <MessageEndpointMappings>
+      <add Messages="Sales.Contract"
+           Endpoint="sales"/>
+      <add Messages="Billing.Contract"
+           Endpoint="billing"/>
+    </MessageEndpointMappings>
+  </UnicastBusConfig>
+  <!--endcode-->
+</configuration>
