+++ Snippets/SqlServer/SqlServer_2.1/SqlServer_2.1.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{2C8CDF08-D497-4FF9-B417-0FE4D3856FCB}</ProjectGuid>
+    <ProjectGuid>{DB7B97B4-3075-4541-A821-1C6FF9AB8671}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>SqlServer_2</RootNamespace>
-    <AssemblyName>SqlServer_2</AssemblyName>
+    <RootNamespace>SqlServer_2._1</RootNamespace>
+    <AssemblyName>SqlServer_2.1</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <TargetFrameworkProfile />
@@ -21,33 +21,53 @@
     <DefineConstants>DEBUG;TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-    <UseVSHostingProcess>false</UseVSHostingProcess>
-    <LangVersion>6</LangVersion>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
+    <DebugType>pdbonly</DebugType>
+    <Optimize>true</Optimize>
+    <OutputPath>bin\Release\</OutputPath>
+    <DefineConstants>TRACE</DefineConstants>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="System" />
     <Reference Include="System.Core" />
+    <Reference Include="System.Xml.Linq" />
+    <Reference Include="System.Data.DataSetExtensions" />
+    <Reference Include="Microsoft.CSharp" />
     <Reference Include="System.Data" />
-  </ItemGroup>
-  <ItemGroup>
-    <Content Include="ConnectionStrings.xml" />
-    <Content Include="ConnectionStringsNamed.xml" />
-    <Content Include="MultiSchema.xml" />
+    <Reference Include="System.Net.Http" />
+    <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ConfigurationSettings.cs" />
-    <Compile Include="NamedConnectionString.cs" />
+    <Compile Include="MultiDb.cs" />
     <Compile Include="MultiSchema.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <Content Include="ExpiredIndex.sql" />
-    <Content Include="MultiDb.config.xml">
-      <SubType>Designer</SubType>
-    </Content>
+    <Compile Include="NamedConnectionString.cs" />
+    <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="UpgradeGuides\MultiInstance.cs" />
+    <Compile Include="UpgradeGuides\Transactions.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="project.json" />
   </ItemGroup>
+  <ItemGroup>
+    <Content Include="ConnectionStrings.xml" />
+    <Content Include="ConnectionStringsNamed.xml" />
+    <Content Include="ExpiredIndex.sql" />
+    <Content Include="MultiDb.config.xml">
+      <SubType>Designer</SubType>
+    </Content>
+    <Content Include="MultiSchema.xml" />
+    <Content Include="UpgradeGuides\ExpiredIndex.sql" />
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
+  <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
+       Other similar extension points exist, see Microsoft.Common.targets.
+  <Target Name="BeforeBuild">
+  </Target>
+  <Target Name="AfterBuild">
+  </Target>
+  -->
 </Project>
