+++ Snippets/SqlServer/SqlServer_2.1/UpgradeGuides/ExpiredIndex.sql
