+++ Snippets/SqlServer/SqlServer_2.1/project.json
@@ -1,6 +1,6 @@
 ﻿{
   "dependencies": {
-    "NServiceBus.SqlServer": "2.0.*"
+    "NServiceBus.SqlServer": "2.1.*"
   },
   "frameworks": {
     "net461": {}
