+++ Snippets/SqlServer/SqlServer_2/ConfigurationSettings.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using NServiceBus;
+﻿using NServiceBus;
 
 class SqlServerConfigurationSettings
 {
@@ -13,27 +12,7 @@ class SqlServerConfigurationSettings
 
         #endregion
     }
-
-    void TimeToWaitBeforeTriggeringCircuitBreaker(BusConfiguration busConfiguration)
-    {
-        #region sqlserver-TimeToWaitBeforeTriggeringCircuitBreaker
-
-        var transport = busConfiguration.UseTransport<SqlServerTransport>();
-        transport.TimeToWaitBeforeTriggeringCircuitBreaker(TimeSpan.FromMinutes(3));
-
-        #endregion
-    }
-
-    void PauseAfterReceiveFailure(BusConfiguration busConfiguration)
-    {
-        #region sqlserver-PauseAfterReceiveFailure
-
-        var transport = busConfiguration.UseTransport<SqlServerTransport>();
-        transport.PauseAfterReceiveFailure(TimeSpan.FromSeconds(15));
-
-        #endregion
-    }
-
+    
     void DisableSecondaries(BusConfiguration busConfiguration)
     {
         #region sqlserver-config-disable-secondaries
