+++ Snippets/SqlServer/SqlServer_3/MultiSchema.xml
@@ -10,20 +10,4 @@
     </MessageEndpointMappings>
   </UnicastBusConfig>
   <!--endcode-->
-  <!--startcode sqlserver-non-standard-schema-messagemapping -->
-  <connectionStrings>
-    <add name="NServiceBus/Transport"
-         connectionString="Data Source=INSTANCE_NAME; Initial Catalog=some_database; Integrated Security=True; Max Pool Size=80"/>
-    <add name="NServiceBus/Transport/Billing"
-         connectionString="Data Source=INSTANCE_NAME; Initial Catalog=some_database; Integrated Security=True; Queue Schema=receiver; Max Pool Size=80"/>
-  </connectionStrings>
-  <UnicastBusConfig>
-    <MessageEndpointMappings>
-      <add Messages="Sales.Contract"
-           Endpoint="sales@salesSchema"/>
-      <add Messages="Billing.Contract"
-           Endpoint="[billing]@[billingSchema]"/>
-    </MessageEndpointMappings>
-  </UnicastBusConfig>
-  <!--endcode-->
 </configuration>
