+++ nservicebus/sqlserver/connection-settings_circuit-breaker_sqlserver_[2.1,3).partial.md
