+++ nservicebus/sqlserver/connection-settings_custom-schema_sqlserver_[2,2.1).partial.md
@@ -1,0 +1,6 @@
+The default schema can be overridden via the connection string, using `Queue Schema` parameter:
+
+snippet:sqlserver-non-standard-schema-connString
+
+snippet:sqlserver-non-standard-schema-connString-xml
+
