+++ nservicebus/sqlserver/connection-settings_custom-schema_sqlserver_[2.1,3).partial.md
