+++ nservicebus/sqlserver/connection-settings_factory_sqlserver_[2.1,).partial.md
