+++ nservicebus/sqlserver/connection-settings_multi-instance_sqlserver_[3,).partial.md
@@ -3,6 +3,8 @@ In [*multi-catalog* and *multi-instance* modes](/nservicebus/sqlserver/deploymen
  * The sending endpoint needs the connection string of the receiving endpoint.
  * The subscribing endpoint needs the connection string of the publishing endpoint, in order to send subscription request.
 
-Connection strings for the remote endpoint can be configured via `app.config` convention
+Connection strings for the remote endpoint can be configured using following API
 
-partial: multiple-appconfig
+snippet:sqlserver-multidb-other-endpoint-connection
+
+NOTE: The `address` parameter passed to the callback above is a transport address. It conforms to the `queue@[schema]` convention, e.g. could be equal to `MultiInstanceSender@[dbo]`.
