--- nservicebus/sqlserver/connection-settings_multiple-connection-push_sqlserver_[,2].partial.md
@@ -1,7 +1,0 @@
-
-### Via the configuration API - Push mode
-
-In the push mode the whole collection of endpoint connection information objects is passed during configuration time.
-
-snippet:sqlserver-multidb-other-endpoint-connection-push
-
