--- nservicebus/sqlserver/connections-settings_multi-schema-config-notes_sqlserver[3,).partial.md
@@ -1,3 +1,0 @@
-NOTE: The table and schema names can be passed either using common convention with square brackets, or without them. 
-
-NOTE: Custom schema need to configured either by app.config **OR** the code api. The two approaches cannot be used at the same time.
